fix: use traces to avoid re-render issues (DX-000)
